first-script/requirements.txt
requests==2.21.0
kafka-python==2.0.2